Move file locations; cleanup
